Flytt utils-funksjon nærmere der den blir brukt, fiks importer i test
