+++ src/frontend/sider/Fagsak/Behandling/sider/Vedtak/Vedtaksperioder/VedtakBegrunnelserContext.test.ts
@@ -1,15 +1,17 @@
 import { addMonths, endOfMonth, startOfMonth } from 'date-fns';
 
-import { BehandlingResultat, BehandlingStatus } from '../../../../../../../typer/behandling';
-import type { IVedtaksperiodeMedBegrunnelser } from '../../../../../../../typer/vedtaksperiode';
-import { Vedtaksperiodetype } from '../../../../../../../typer/vedtaksperiode';
-import { dagensDato, dateTilIsoDatoString } from '../../../../../../../utils/dato';
+import { filtrerOgSorterPerioderMedBegrunnelseBehov } from './utils';
+import { BehandlingResultat, BehandlingStatus } from '../../../../../../typer/behandling';
+import {
+    Vedtaksperiodetype,
+    type IVedtaksperiodeMedBegrunnelser,
+} from '../../../../../../typer/vedtaksperiode';
+import { dagensDato, dateTilIsoDatoString } from '../../../../../../utils/dato';
 import {
     mockAvslagsperiode,
     mockOpphørsperiode,
     mockUtbetalingsperiode,
-} from '../../../../../../../utils/test/vedtak/vedtaksperiode.mock';
-import { filtrerOgSorterPerioderMedBegrunnelseBehov } from '../../../../../../../utils/vedtakUtils';
+} from '../../../../../../utils/test/vedtak/vedtaksperiode.mock';
 
 describe('VedtaksBegrunnelserContext', () => {
     describe('Test filtrerOgSorterPerioderMedBegrunnelseBehov', () => {
