+++ src/frontend/sider/Fagsak/Behandling/sider/Vedtak/Vedtaksperioder/utils.ts
@@ -1,6 +1,16 @@
+import {
+    addMonths,
+    differenceInMilliseconds,
+    isAfter,
+    isBefore,
+    isSameMonth,
+    startOfMonth,
+} from 'date-fns';
+
 import type { GroupBase, OptionType } from '@navikt/familie-form-elements';
 import { RessursStatus, type Ressurs } from '@navikt/familie-typer';
 
+import { BehandlingResultat, BehandlingStatus } from '../../../../../../typer/behandling';
 import {
     BegrunnelseType,
     begrunnelseTyper,
@@ -13,6 +23,13 @@ import {
     type IVedtaksperiodeMedBegrunnelser,
 } from '../../../../../../typer/vedtaksperiode';
 import type { VedtaksbegrunnelseTekster } from '../../../../../../typer/vilkår';
+import {
+    dagensDato,
+    isoStringTilDate,
+    isoStringTilDateMedFallback,
+    tidenesMorgen,
+    type IsoDatoString,
+} from '../../../../../../utils/dato';
 
 export function grupperteBegrunnelser(
     vedtaksperiodeMedBegrunnelser: IVedtaksperiodeMedBegrunnelser,
@@ -115,4 +132,77 @@ const hentLabelForOption = (
                   restVedtakBegrunnelseTilknyttetVilkår.id === begrunnelse
           )?.navn ?? '')
         : '';
+};
+
+export const filtrerOgSorterPerioderMedBegrunnelseBehov = (
+    vedtaksperioder: IVedtaksperiodeMedBegrunnelser[],
+    behandlingResultat: BehandlingResultat,
+    behandlingStatus: BehandlingStatus,
+    sisteVedtaksperiodeVisningDato: IsoDatoString | undefined,
+    skalAlltidViseAlleVedtaksperioder: boolean
+): IVedtaksperiodeMedBegrunnelser[] => {
+    const sorterteOgFiltrertePerioder = vedtaksperioder
+        .slice()
+        .sort((a, b) =>
+            differenceInMilliseconds(
+                isoStringTilDateMedFallback({ isoString: a.fom, fallbackDate: tidenesMorgen }),
+                isoStringTilDateMedFallback({ isoString: b.fom, fallbackDate: tidenesMorgen })
+            )
+        )
+        .filter((vedtaksperiode: IVedtaksperiodeMedBegrunnelser) => {
+            if (behandlingStatus === BehandlingStatus.AVSLUTTET) {
+                return harPeriodeBegrunnelse(vedtaksperiode);
+            } else if (skalAlltidViseAlleVedtaksperioder) {
+                return true;
+            } else {
+                return (
+                    (sisteVedtaksperiodeVisningDato &&
+                        erPeriodeMindreEllerLikEnnSisteVedtaksperiodeVisningDato(
+                            sisteVedtaksperiodeVisningDato,
+                            vedtaksperiode.fom
+                        )) ||
+                    erPeriode2MndFramITidEllerMindre(vedtaksperiode.fom)
+                );
+            }
+        });
+
+    if (
+        behandlingResultat === BehandlingResultat.OPPHØRT ||
+        behandlingResultat === BehandlingResultat.FORTSATT_OPPHØRT
+    ) {
+        return hentSisteOpphørsperiode(sorterteOgFiltrertePerioder);
+    } else {
+        return sorterteOgFiltrertePerioder;
+    }
+};
+
+const erPeriodeMindreEllerLikEnnSisteVedtaksperiodeVisningDato = (
+    sisteVedtaksperiodeVisningDato: string,
+    periode: string | undefined
+) => {
+    return isAfter(
+        isoStringTilDate(sisteVedtaksperiodeVisningDato),
+        isoStringTilDateMedFallback({ isoString: periode, fallbackDate: tidenesMorgen })
+    );
+};
+
+const erPeriode2MndFramITidEllerMindre = (periodeFom: string | undefined) => {
+    const fom = isoStringTilDateMedFallback({ isoString: periodeFom, fallbackDate: tidenesMorgen });
+    const toMånederFremITid = addMonths(startOfMonth(dagensDato), 2);
+
+    return isBefore(fom, toMånederFremITid) || isSameMonth(fom, toMånederFremITid);
+};
+
+const harPeriodeBegrunnelse = (vedtaksperiode: IVedtaksperiodeMedBegrunnelser) => {
+    return !!vedtaksperiode.begrunnelser.length || !!vedtaksperiode.fritekster.length;
+};
+
+const hentSisteOpphørsperiode = (sortertePerioder: IVedtaksperiodeMedBegrunnelser[]) => {
+    const sorterteOgFiltrerteOpphørsperioder = sortertePerioder.filter(
+        (vedtaksperiode: IVedtaksperiodeMedBegrunnelser) =>
+            vedtaksperiode.type === Vedtaksperiodetype.OPPHØR
+    );
+    const sisteOpphørsPeriode =
+        sorterteOgFiltrerteOpphørsperioder[sorterteOgFiltrerteOpphørsperioder.length - 1];
+    return sisteOpphørsPeriode ? [sisteOpphørsPeriode] : [];
 };
