+++ src/frontend/utils/vedtakUtils.ts
@@ -1,13 +1,4 @@
 import {
-    addMonths,
-    differenceInMilliseconds,
-    isAfter,
-    isBefore,
-    isSameMonth,
-    startOfMonth,
-} from 'date-fns';
-
-import {
     ABorderDanger,
     ABorderDefault,
     ABorderSubtle,
@@ -22,87 +13,10 @@ import {
 import type { Ressurs } from '@navikt/familie-typer';
 import { RessursStatus } from '@navikt/familie-typer';
 
-import type { IsoDatoString } from './dato';
-import { dagensDato, isoStringTilDate, isoStringTilDateMedFallback, tidenesMorgen } from './dato';
-import { BehandlingResultat, BehandlingStatus } from '../typer/behandling';
+import { BehandlingResultat } from '../typer/behandling';
 import type { IRestBegrunnelseTilknyttetVilkår, Begrunnelse } from '../typer/vedtak';
 import { BegrunnelseType } from '../typer/vedtak';
-import type { IVedtaksperiodeMedBegrunnelser } from '../typer/vedtaksperiode';
-import { Vedtaksperiodetype } from '../typer/vedtaksperiode';
 import type { VedtaksbegrunnelseTekster } from '../typer/vilkår';
-
-export const filtrerOgSorterPerioderMedBegrunnelseBehov = (
-    vedtaksperioder: IVedtaksperiodeMedBegrunnelser[],
-    behandlingResultat: BehandlingResultat,
-    behandlingStatus: BehandlingStatus,
-    sisteVedtaksperiodeVisningDato: IsoDatoString | undefined,
-    skalAlltidViseAlleVedtaksperioder: boolean
-): IVedtaksperiodeMedBegrunnelser[] => {
-    const sorterteOgFiltrertePerioder = vedtaksperioder
-        .slice()
-        .sort((a, b) =>
-            differenceInMilliseconds(
-                isoStringTilDateMedFallback({ isoString: a.fom, fallbackDate: tidenesMorgen }),
-                isoStringTilDateMedFallback({ isoString: b.fom, fallbackDate: tidenesMorgen })
-            )
-        )
-        .filter((vedtaksperiode: IVedtaksperiodeMedBegrunnelser) => {
-            if (behandlingStatus === BehandlingStatus.AVSLUTTET) {
-                return harPeriodeBegrunnelse(vedtaksperiode);
-            } else if (skalAlltidViseAlleVedtaksperioder) {
-                return true;
-            } else {
-                return (
-                    (sisteVedtaksperiodeVisningDato &&
-                        erPeriodeMindreEllerLikEnnSisteVedtaksperiodeVisningDato(
-                            sisteVedtaksperiodeVisningDato,
-                            vedtaksperiode.fom
-                        )) ||
-                    erPeriode2MndFramITidEllerMindre(vedtaksperiode.fom)
-                );
-            }
-        });
-
-    if (
-        behandlingResultat === BehandlingResultat.OPPHØRT ||
-        behandlingResultat === BehandlingResultat.FORTSATT_OPPHØRT
-    ) {
-        return hentSisteOpphørsperiode(sorterteOgFiltrertePerioder);
-    } else {
-        return sorterteOgFiltrertePerioder;
-    }
-};
-
-const erPeriodeMindreEllerLikEnnSisteVedtaksperiodeVisningDato = (
-    sisteVedtaksperiodeVisningDato: string,
-    periode: string | undefined
-) => {
-    return isAfter(
-        isoStringTilDate(sisteVedtaksperiodeVisningDato),
-        isoStringTilDateMedFallback({ isoString: periode, fallbackDate: tidenesMorgen })
-    );
-};
-
-const erPeriode2MndFramITidEllerMindre = (periodeFom: string | undefined) => {
-    const fom = isoStringTilDateMedFallback({ isoString: periodeFom, fallbackDate: tidenesMorgen });
-    const toMånederFremITid = addMonths(startOfMonth(dagensDato), 2);
-
-    return isBefore(fom, toMånederFremITid) || isSameMonth(fom, toMånederFremITid);
-};
-
-const harPeriodeBegrunnelse = (vedtaksperiode: IVedtaksperiodeMedBegrunnelser) => {
-    return !!vedtaksperiode.begrunnelser.length || !!vedtaksperiode.fritekster.length;
-};
-
-const hentSisteOpphørsperiode = (sortertePerioder: IVedtaksperiodeMedBegrunnelser[]) => {
-    const sorterteOgFiltrerteOpphørsperioder = sortertePerioder.filter(
-        (vedtaksperiode: IVedtaksperiodeMedBegrunnelser) =>
-            vedtaksperiode.type === Vedtaksperiodetype.OPPHØR
-    );
-    const sisteOpphørsPeriode =
-        sorterteOgFiltrerteOpphørsperioder[sorterteOgFiltrerteOpphørsperioder.length - 1];
-    return sisteOpphørsPeriode ? [sisteOpphørsPeriode] : [];
-};
 
 export const finnBegrunnelseType = (
     vilkårBegrunnelser: Ressurs<VedtaksbegrunnelseTekster>,
